updated dependencies and got the web part to work
