+++ build.sbt
@@ -1,4 +1,9 @@
+import xerial.sbt.Pack._
+import sbt._
+import Keys._
+
 name := "activator-akka-spray"
+
 
 version := "1.0"
 
@@ -8,18 +13,22 @@ resolvers += "spray repo" at "http://repo.spray.io"
 
 resolvers += "spray nightlies" at "http://nightlies.spray.io"
 
-libraryDependencies ++= Seq(
-  "com.typesafe.akka"  %% "akka-actor"       % "2.2.0",
-  "com.typesafe.akka"  %% "akka-slf4j"       % "2.2.0",
-  "ch.qos.logback"      % "logback-classic"  % "1.0.13",
-  "io.spray"            % "spray-can"        % "1.2-20130712",
-  "io.spray"            % "spray-routing"    % "1.2-20130712",
-  "io.spray"           %% "spray-json"       % "1.2.3",
-  "org.specs2"         %% "specs2"           % "1.14"         % "test",
-  "io.spray"            % "spray-testkit"    % "1.2-20130712" % "test",
-  "com.typesafe.akka"  %% "akka-testkit"     % "2.2.0"        % "test",
-  "com.novocode"        % "junit-interface"  % "0.7"          % "test->default"
-)
+libraryDependencies ++= {
+  val akkaVersion  = "2.2.3"
+  val sprayVersion = "1.2.0"
+  Seq(
+    "com.typesafe.akka" %% "akka-actor" % akkaVersion,
+    "com.typesafe.akka" %% "akka-slf4j" % akkaVersion,
+    "ch.qos.logback" % "logback-classic" % "1.0.13",
+    "io.spray" % "spray-can" % sprayVersion,
+    "io.spray" % "spray-routing" % sprayVersion,
+    "io.spray" %% "spray-json" % "1.2.5",
+    "org.specs2" %% "specs2" % "1.14" % "test",
+    "io.spray" % "spray-testkit" % sprayVersion % "test",
+    "com.typesafe.akka" %% "akka-testkit" % akkaVersion % "test",
+    "com.novocode" % "junit-interface" % "0.7" % "test->default"
+  )
+}
 
 scalacOptions ++= Seq(
   "-unchecked",
@@ -32,3 +41,13 @@ scalacOptions ++= Seq(
 )
 
 testOptions += Tests.Argument(TestFrameworks.JUnit, "-v")
+
+unmanagedResourceDirectories in Compile <++= baseDirectory {
+  base => Seq(base / "src/main/angular")
+}
+
+packSettings
+
+packMain := Map("rest" -> "Rest")
+
+crossPaths := false
