+++ project/plugins.sbt
@@ -1,0 +1,4 @@
+
+addSbtPlugin("org.xerial.sbt" % "sbt-pack" % "0.4.1")
+
+
