+++ src/main/angular/root/css/jquery.fileupload-ui-noscript.css
