+++ src/main/angular/root/css/style.css
