+++ src/main/angular/root/index-ng-upload.html
