+++ src/main/angular/root/js/jquery.fileupload-angular.js
@@ -19,10 +19,10 @@
         define([
             'jquery',
             'angular',
-            './jquery.fileupload-image',
-            './jquery.fileupload-audio',
-            './jquery.fileupload-video',
-            './jquery.fileupload-validate'
+            './jquery.fileupload-image.js',
+            './jquery.fileupload-audio.js',
+            './jquery.fileupload-video.js',
+            './jquery.fileupload-validate.js'
         ], factory);
     } else {
         factory();
