+++ src/main/angular/root/js/jquery.fileupload.js
