+++ src/main/angular/root/js/ng-upload.js
