--- src/main/scala/Cli.scala
@@ -1,3 +1,0 @@
-import core.{BootedCore, CoreActors}
-
-object Cli extends App with BootedCore with CoreActors
