+++ src/main/scala/Rest.scala
@@ -1,5 +1,10 @@
+import akka.actor.{ActorSystem, ActorRefFactory}
 import api.Api
-import core.{BootedCore, CoreActors}
-import web.Web
+import core.{Core, BootedCore, CoreActors}
+import web.{StaticResources, Web}
 
-object Rest extends App with BootedCore with CoreActors with Api with Web
+object Rest extends App with BootedCore with Core with CoreActors with Api  with StaticResources  {
+
+
+}
+
