+++ src/main/scala/api/Api.scala
@@ -1,24 +1,22 @@
 package api
 
 import core.{CoreActors, Core}
-import akka.actor.Props
-import spray.routing.RouteConcatenation
-
+import akka.actor.{ActorSystem, Props}
+import spray.routing.{HttpService, RouteConcatenation}
+import scala.concurrent.ExecutionContext
+import ExecutionContext.Implicits.global
 /**
  * The REST API layer. It exposes the REST services, but does not provide any
  * web server interface.<br/>
  * Notice that it requires to be mixed in with ``core.CoreActors``, which provides access
  * to the top-level actors that make up the system.
  */
-trait Api extends RouteConcatenation {
-  this: CoreActors with Core =>
+trait Api extends HttpService with CoreActors with Core {
 
-  private implicit val _ = system.dispatcher
-
+  //protected implicit val system : ActorSystem
   val routes =
     new RegistrationService(registration).route ~
     new MessengerService(messenger).route
 
-  val rootService = system.actorOf(Props(new RoutedHttpService(routes)))
 
 }
