+++ src/main/scala/api/services.scala
@@ -3,8 +3,8 @@ package api
 import spray.http.StatusCodes._
 import spray.http._
 import spray.routing._
-import directives.{CompletionMagnet, RouteDirectives}
-import spray.util.{SprayActorLogging, LoggingContext}
+import directives.{ RouteDirectives}
+import spray.util.{ LoggingContext}
 import util.control.NonFatal
 import spray.httpx.marshalling.Marshaller
 import spray.http.HttpHeaders.RawHeader
@@ -66,7 +66,7 @@ trait FailureHandling {
  *
  * @param route the (concatenated) route
  */
-class RoutedHttpService(route: Route) extends Actor with HttpService with SprayActorLogging {
+class RoutedHttpService(route: Route) extends Actor with HttpService with ActorLogging {
 
   implicit def actorRefFactory = context
 
@@ -94,7 +94,7 @@ class RoutedHttpService(route: Route) extends Actor with HttpService with SprayA
 trait CrossLocationRouteDirectives extends RouteDirectives {
 
   implicit def fromObjectCross[T : Marshaller](origin: String)(obj: T) =
-    new CompletionMagnet {
+     {
       def route: StandardRoute = new CompletionRoute(OK,
         RawHeader("Access-Control-Allow-Origin", origin) :: Nil, obj)
     }
