+++ src/main/scala/core/core.scala
@@ -1,6 +1,10 @@
 package core
 
-import akka.actor.{Props, ActorSystem}
+import akka.actor.{Props, ActorRefFactory, ActorSystem}
+import api.{Api, RoutedHttpService}
+import akka.io.IO
+import spray.can.Http
+import web.StaticResources
 
 /**
  * Core is type containing the ``system: ActorSystem`` member. This enables us to use it in our
@@ -8,7 +12,7 @@ import akka.actor.{Props, ActorSystem}
  */
 trait Core {
 
-  implicit def system: ActorSystem
+  protected implicit def system: ActorSystem
 
 }
 
@@ -16,12 +20,17 @@ trait Core {
  * This trait implements ``Core`` by starting the required ``ActorSystem`` and registering the
  * termination handler to stop the system when the JVM exits.
  */
-trait BootedCore extends Core {
+trait BootedCore extends Core with Api with StaticResources {
+  def system: ActorSystem = ActorSystem("activator-akka-spray")
+  def actorRefFactory: ActorRefFactory = system
+  val rootService = system.actorOf(Props(new RoutedHttpService(routes ~ staticResources )))
+
+  IO(Http)(system) ! Http.Bind(rootService, "0.0.0.0", port = 8080)
 
   /**
    * Construct the ActorSystem we will use in our application
    */
-  implicit lazy val system = ActorSystem("akka-spray")
+  //protected implicit  val system : ActorSystem
 
   /**
    * Ensure that the constructed ActorSystem is shut down when the JVM shuts down
